Rename test files to lowercase
